feat: conditionally render the new MoreInfo feedback content
- Split the new UI into it's own component
- Conditionally render the new UI when the feature flag is enabled
